Added GetAuthorizationStatus
Allows us to determine the current authorization status of Geo location, so that we can cater for alternative UI regarding the UX. e.g. iOS - When the user hasn't ever set a permission, the status will be undetermined, so we can present a different UI message than that of the status being unauthorized.