make static selectors internal
These properties are used internally for caching purposes, and should not
be exposed in the public API, nor show up in docs, etc.

There are instance properties available for Uno programmers. The static
selector properties are unlikely to be used directly by external code, and
because we are working on 2.0 it is OK to clean-up our public API a little.