drop obsolete legacy code
This removes unused obsolete legacy code before our next major release.

Also makes the NavigationPageProxy class internal as it was supposed to
be in the first place.

This seems unlikely to affect third-party code out in the wild, but if
anyone runs into trouble: Feel free to open issues on GitHub or ping me
directly on Slack, and I will help you migrate. Thanks!