Revert back to old Debug.Log()
This reverts parts of commit 61f59e7c64045e264d927586b600557af56eb84f,
and fixes compatibility with older beta-releases of fuse X (before
2.0.0-beta.7).

The problem is that fuse X ships with a prebuilt version of UnoCore,
and older releases don't contain the new Log-class, causing run-time
errors when using the latest beta-release of Fuselibs.

So let's revert back for now, and switch to the new Log-class once we
are done with beta-releases.