Android: don't add BACKGROUND_LOCATION to manifest
Apps with this permission declared in their manifest are rejected when
submitting to Google Play, unless providing a video demonstrating how
this functionality is used in the app.

The problem with some apps is that they aren't really using this feature
so there's nothing to demonstrate, and the only way to publish currently
is to manually edit AndroidManifest.xml, remove the permission, re-run
build.sh, and finally submit the modified AAB/APK.

This patch addresses the problem by not adding this permission
automatically. Apps that really need it will be required to add it
manually. Please report issues if this becomes a problem for anyone.

Closes #1402