WebView (Android): download ability
This makes it possible to download files when using WebView on Android.

Two new Uno events are added to the WebView class in order to process
downloaded files:

  * BeginDownload(WebView sender, DownloadEventArgs args)
  * FileDownloaded(WebView sender, DownloadEventArgs args)

It might be possible to access the events from JavaScript, but this has
not yet been explored.

Downloaded files can also be found using a File Manager app.